tidy up specs
