+++ spec/integration/basic/list/rails_admin_basic_list_spec.rb
@@ -2,18 +2,6 @@
 
 require 'spec_helper'
 
-def checkbox_filter_selector(association_name, model)
-  "input[name='f[#{association_name}][1][v][]'][value='#{model.id}']"
-end
-
-def text_filter_selector(association_name)
-  "input[name='f[#{association_name}][1][v]'][type='text']"
-end
-
-def select_filter_selector(association_name)
-  "select[name='f[#{association_name}][1][v]']"
-end
-
 describe 'RailsAdmin Basic List', type: :request do
   subject { page }
